Переход. Поправлен тип поля price модели product

+++ catalog/models.py
@@ -112,7 +112,11 @@ class Product(CategoryProductsAbstractModel):
 
     folder_img = os.path.join('catalog', 'products')
 
-    price = models.PositiveIntegerField(verbose_name='Цена',)
+    price = models.DecimalField(
+        verbose_name='Цена',
+        max_digits=10,
+        decimal_places=2
+    )
     subcategory = models.ForeignKey(
         SubCategory,
         verbose_name='Подкатегория',
